Avoid loading Psych 4.0 on unsupported versions
There's an issue where the latest rdoc is pulling in Psych 4.0.  This causes issues when running in CI on Ruby 2.5.

This PR addresses the issue by not loading rdoc for Rails versions that currently run against Ruby 2.5.  The Rakefile is updated so that it can still run when rdoc is unavailable